Fix number of symbols transmited in each iteration to 140
Now each iteration corresponds to a transmission of an LTE frame.
This change is motivated to the easy of mapping of the LTE grid if
data is generated on a per-frame basis. Since 140 OFDM symbols might
lead to an excessive quantity of printings in the console, all
printings can now be turned off.